ai_cbt_trainer_v1.0/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
python-docx==1.1.0
